pyproject.toml
[project]
name = "easyopenai"
version = "0.1.0"
description = "Async OpenAI-compatible client with multi-provider scheduling, rate limiting, retries, and reasoning-model support."
readme = "README.md"
authors = [
    { name = "jubgjf", email = "guanjiannan@outlook.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiolimiter>=1.2.1",
    "loguru>=0.7.3",
    "openai>=2.32.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "tenacity>=9.1.4",
]

[project.scripts]
easyopenai = "easyopenai:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "respx>=0.23.1",
]
